% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_submission.R
\name{format_rt_forecast}
\alias{format_rt_forecast}
\title{Format Rt forecasts to Submission Format}
\usage{
format_rt_forecast(
  loc = NULL,
  loc_name = NULL,
  forecast_date = NULL,
  forecast_adjustment = 16,
  version = "1.0"
)
}
\arguments{
\item{loc}{path to folder}

\item{loc_name}{name of region}

\item{forecast_date}{date forecast is made}

\item{forecast_adjustment}{shift forecasts by a certain number of dates}
}
\value{
data.frame with forecats in submission format
}
\description{
Format Rt forecasts to Submission Format
}
\details{
loads Rt forecasts and converts them to the correct format
}
