% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_submission.R
\name{format_timeseries_forecast}
\alias{format_timeseries_forecast}
\title{Format timeseries forecasts to Submission Format}
\usage{
format_timeseries_forecast(model_type)
}
\arguments{
\item{model_type}{can be either "deaths-only" or "deaths-on-cases"}
}
\value{
data.frame with forecats in submission format
}
\description{
Format timeseries forecasts to Submission Format
}
\details{
loads timeseries forecasts and converts them to the correct format
}
\examples{

deaths_on_cases <- format_timeseries_forecast(model_type = "deaths-on-cases")
readr::write_csv(deaths_on_cases, here::here("timeseries-forecast", "deaths-on-cases", 
                                             "submission-files",
                                             "latest-weekly-deaths-on-cases.csv"))

deaths_only <- format_timeseries_forecast(model_type = "deaths-only")
readr::write_csv(deaths_only, here::here("timeseries-forecast", "deaths-only", 
                                         "submission-files",
                                         "latest-weekly-deaths-only.csv"))
}
