% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stored_forecasts.R
\name{get_raw_case_forecasts}
\alias{get_raw_case_forecasts}
\title{Load in Raw Case Forecasts from Rt model}
\usage{
get_raw_case_forecasts(results_dir = NULL, date = NULL)
}
\arguments{
\item{results_dir}{directory to the results}

\item{date}{date for which to get the raw case forecasts}
}
\value{
data.frame with raw case forecasts
}
\description{
Load in Raw Case Forecasts from Rt model
}
\details{
Loads raw case forecasts made by the Rt model for a single date. 
Function is adapted from EpiNow::get_timeseries()
}
\examples{

}
