% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stored_forecasts.R
\name{get_true_and_forecast}
\alias{get_true_and_forecast}
\title{Load Observed Deaths Together With Past Forecasts}
\usage{
get_true_and_forecast(forecast_dates = NULL)
}
\arguments{
\item{forecast_dates}{forecast_dates specify the forecast dates for which to load corresponding
forecasts. If \code{forecast_dates = NULL} all dates for which there are
US forecasts available will be loaded}
}
\value{
data.frame with obsered deaths
}
\description{
Load Observed Deaths Together With Past Forecasts
}
\details{
Loads death data from file and the forecasts made by the different models
}
\examples{
df <- get_true_and_forecast()

data.table::fwrite(df, here::here("evaluation", "observed_vs_forecast", "rt-forecast-vs-obs.csv"))
}
