% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_data.R
\name{get_us_cases}
\alias{get_us_cases}
\title{Get US Case Data}
\usage{
get_us_cases(data = c("cumulative", "daily"))
}
\arguments{
\item{data}{type. Can be either "daily" or "cumulative"}
}
\value{
data.frame with Deaths Data
}
\description{
Get US Case Data
}
\details{
Get & reshape JHU data
Source: https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data
}
