% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_data.R
\name{get_us_deaths}
\alias{get_us_deaths}
\title{Get US Death Data}
\usage{
get_us_deaths(data = c("cumulative", "daily"))
}
\value{
data.frame with Deaths Data
}
\description{
Get US Death Data
}
\details{
Get & reshape JHU data
Source: https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data
}
