% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stored_forecasts.R
\name{load_all_rt_forecasts}
\alias{load_all_rt_forecasts}
\title{Load in Complete Set of Raw Case Forecasts from Rt model}
\usage{
load_all_rt_forecasts(
  forecast_dates = NULL,
  forecast_adjustment = 16,
  weekly = FALSE
)
}
\arguments{
\item{forecast_dates}{specify the forecast dates for which to load corresponding
forecasts. If \code{forecast_dates = NULL} all dates for which there are
US forecasts available will be loaded}

\item{forecast_adjustment}{shift dates for the forecasts by a certain amount}

\item{weekly}{return weekly instead of daily data}
}
\value{
data.frame with all Rt model forecasts
}
\description{
Load in Complete Set of Raw Case Forecasts from Rt model
}
\details{
loads in national and subnational case forecasts from the Rt model and 
formats them
}
\examples{

}
