% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_data.R
\name{load_observed_deaths}
\alias{load_observed_deaths}
\title{Load Observed Deaths}
\usage{
load_observed_deaths(weekly = FALSE, cumulative = FALSE)
}
\arguments{
\item{weekly}{return data in a weekly format}

\item{cumulative}{return cumulative data instead of incidence data}
}
\value{
data.frame with obsered deaths
}
\description{
Load Observed Deaths
}
\details{
Loads death data from file and returns a data.table
}
