% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toy_data.R
\name{make_toy_data}
\alias{make_toy_data}
\title{Create Toy Data}
\usage{
make_toy_data(
  forecast_dates = NULL,
  y_true = NULL,
  horizons = 1:10,
  regions = c("US", "Alabama"),
  samples = 5,
  models = c("model1", "model2")
)
}
\arguments{
\item{forecast_dates}{dates for which to create forecasts}

\item{y_true}{allows to pass a vector with true values that can be used to 
replace the randomly generated ones. Useful if you want true values to match
across different toy data sets}

\item{horizons}{horizons to "forecast"}

\item{regions}{names of regions}

\item{samples}{number of samples}

\item{models}{names for the toy models}
}
\value{
data.frame with toy predictions
}
\description{
Create Toy Data
}
\details{
Creates toy data with some predictions vs. true values
}
\examples{
# make toy 1
toy1 <- make_toy_data(samples = 500, 
                      models = "model1")
data.table::fwrite(toy1, here::here("evaluation", "observed_vs_forecast", "toy1-forecast-vs-obs.csv"))

# make toy 2
true_values <- toy1$true_values
toy2 <- make_toy_data(samples = 500, 
                      models = "model2", 
                      y_true = true_values)
#write toy data to file
data.table::fwrite(toy2, here::here("evaluation", "observed_vs_forecast", "toy2-forecast-vs-obs.csv"))

}
